“another”

+++ backend/controllers/cards.js
@@ -6,6 +6,8 @@ const Card = require('../models/cards');
 const ERROR_CODE_400 = require('../errors/error400');
 // Forbidden
 const ERROR_CODE_403 = require('../errors/error403');
+// Not Found
+const ERROR_CODE_404 = require('../errors/error404');
 
 const opts = {
   new: true,
@@ -16,6 +18,7 @@ module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => {
       if (!cards) {
+        res.send({ data: [] });
         throw new ERROR_CODE_400('Переданы некорректные данные при создании карточки.');
       }
       res.send({ data: cards });
@@ -29,27 +32,42 @@ module.exports.createCard = (req, res, next) => {
   Card.create({ name, link, owner })
     .then((card) => {
       if (!card) {
-        throw new ERROR_CODE_400('Переданы некорректные данные при создании карточки.');
+        throw new ERROR_CODE_404('Карточка не найдена, проверьте корректность запроса.');
       }
       res.send({ data: card });
       // Если захочу получать просто карточку, а не объект, то data : надо будет удалить
     })
-    .catch(next);
+    .catch((err) => {
+      if (err.name === 'ValidationError') {
+        next(new ERROR_CODE_400('Карточка не создана, проверьте корректность запроса.'));
+      } else {
+        next(err);
+      }
+    });
 };
 
 module.exports.deleteCard = (req, res, next) => {
   Card.findById(req.params.cardId)
     .then((card) => {
+      if (!card) {
+        throw new ERROR_CODE_404('Карточка не найдена, проверьте корректность запроса.');
+      }
       const owner = card.owner.toString();
       if (req.user._id === owner) {
         Card.deleteOne(card)
           .then(() => {
             res.send(card);
-          })
-          .catch(next);
+          });
+        throw new ERROR_CODE_403('Вы не можете удалить чужую карточку.');
       }
     })
-    .catch(next);
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        next(new ERROR_CODE_400('Карточка не удалена, проверьте корректность запроса.'));
+      } else {
+        next(err);
+      }
+    });
 };
 
 module.exports.likeCard = (req, res, next) => {
@@ -60,11 +78,17 @@ module.exports.likeCard = (req, res, next) => {
   )
     .then((card) => {
       if (!card) {
-        throw new ERROR_CODE_400('Переданы некорректные данные при создании карточки.');
+        throw new ERROR_CODE_404('Карточка не найдена, проверьте корректность запроса.');
       }
       res.send({ data: card });
     })
-    .catch(next);
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        next(new ERROR_CODE_400('Лайк не добавлен, проверьте корректность запроса.'));
+      } else {
+        next(err);
+      }
+    });
 };
 
 module.exports.dislikeCard = (req, res, next) => {
@@ -75,9 +99,15 @@ module.exports.dislikeCard = (req, res, next) => {
   )
     .then((card) => {
       if (!card) {
-        throw new ERROR_CODE_403('Переданы некорректные данные при создании карточки.');
+        throw new ERROR_CODE_404('Карточка не найдена, проверьте корректность запроса.');
       }
       res.send({ data: card });
     })
-    .catch(next);
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        next(new ERROR_CODE_400('Отмена лайка не исполнена, проверьте корректность запроса.'));
+      } else {
+        next(err);
+      }
+    });
 };
